requirements.txt
Flask>=3.0.0
playwright>=1.42.0
beautifulsoup4>=4.12.3
PyYAML>=6.0.1
python-dotenv>=1.0.1
